# api/openapi.yaml
openapi: 3.1.0
info:
  title: ZATRANO API
  version: "1.0.0"
  description: |
    Clean/Hexagonal mimariyle, CQRS ve DDD prensipleriyle inşa edildi. Tüm hatalar RFC 7807 (Problem Details) formatında döner.
    Başarı yanıtları ve e-posta şablonları Accept-Language / ?lang ile yerelleştirilir.

    **Idempotency:** `POST`/`PUT`/`PATCH`/`DELETE` isteklerinde `Idempotency-Key` başlığı gönderilirse
    aynı anahtarla yapılan tekrar istekler ilk yanıtın aynısını döner (çift işlem engeli).
    Web karşılığı: GoUI session/component event koruması ve aynı application use-case idempotency kurallarıdır.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Yerel geliştirme
...
paths:
...
  /api/v1/ws:
    get:
      tags: [realtime]
      summary: Genel canlı olay WebSocket
      description: |
        Upgrade: `Connection: Upgrade` + `Upgrade: websocket`.

        Kimlik (sırayla):
        - `Authorization: Bearer <access>`
        - `?access_token=` veya `?token=`
        - `access_token` cookie (panel same-origin)

        Olaylar JSON text frame:
        ```json
        {"type":"inbox.updated","unread_count":3}
        ```
        Bağlantı açılışında güncel `inbox.updated` gönderilir. Yalnızca o kullanıcıya push edilir.
      security:
        - bearerAuth: []
      responses:
        "101":
          description: Switching Protocols
        "401":
          description: Token yok/geçersiz
        "426":
          description: Upgrade Required
  /api/v1/contact:
    post:
      tags: [public]
      summary: İletişim formu gönderimi
      description: |
        Mesajı veritabanına kaydeder ve CONTACT_RECIPIENT_EMAIL adresine
        kalıcı outbox üzerinden e-posta kuyruğuna alır. Turnstile (yapılandırılmışsa) zorunludur.
      parameters:
        - in: header
          name: X-Turnstile-Token
          schema: { type: string }
          description: Cloudflare Turnstile yanıt token'ı (opsiyonel body alanı da kabul edilir)
        - in: header
          name: Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, message]
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                email: { type: string, format: email }
                message: { type: string, minLength: 5, maxLength: 2000 }
                locale: { type: string, example: tr }
      responses:
        "201":
          description: Mesaj alındı
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  message: { type: string }
        "400": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /api/v1/contacts:
    get:
      tags: [contacts]
      summary: İletişim mesajı listesi
      description: |
        Web karşılığı: GET /dashboard/contacts. İzin: contacts:list
      security: [{ bearerAuth: [] }]
      parameters:
        - name: unread
          in: query
          schema: { type: string, enum: ["1", "true"] }
          description: Yalnızca okunmamış mesajlar
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, enum: [100, 500, 1000], default: 100 }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        "200": { description: Sayfalanmış iletişim mesajı listesi }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/contacts/{id}:
    get:
      tags: [contacts]
      summary: İletişim mesajı detayı
      description: |
        Web karşılığı: GET /dashboard/contacts/{id}. İzin: contacts:list
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: İletişim mesajı }
        "404": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/contacts/{id}/read:
    post:
      tags: [contacts]
      summary: Mesajı okundu işaretle
      description: |
        Web karşılığı: detay sayfasında okundu işaretle. İzin: contacts:list
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: Idempotency-Key
          schema: { type: string }
      responses:
        "200": { description: Güncellenmiş mesaj }
        "404": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Giriş yap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token çifti
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401":
          description: Geçersiz kimlik bilgileri veya MFA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/Problem" }
                  - { $ref: "#/components/schemas/MFARequiredResponse" }
        "429": { $ref: "#/components/responses/Problem" }
...
  /api/v1/auth/forgot-password:
    post:
      tags: [auth]
      summary: Şifre sıfırlama bağlantısı iste (numaralandırmaya karşı jenerik yanıt)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /api/v1/auth/reset-password:
    post:
      tags: [auth]
      summary: Token ile yeni şifre belirle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/verify-email:
    post:
      tags: [auth]
      summary: E-posta adresini doğrula
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/resend-verification:
    post:
      tags: [auth]
      summary: Doğrulama e-postasını yeniden gönder (jenerik yanıt)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResendVerificationRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      summary: Şifre değiştir (mevcut şifreyle; tüm oturumlar iptal edilir)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/mfa/setup:
    post:
      tags: [auth]
      summary: MFA kurulumunu başlat (secret + otpauth URI)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Kurulum bilgisi (yerelleştirilmiş mesaj + veri)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFASetupResponse" }
  /api/v1/auth/mfa/enable:
    post:
      tags: [auth]
      summary: MFA etkinleştir (kod doğrula)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MFACodeRequest" }
      responses:
        "200":
          description: Başarı mesajı ve yalnızca bir kez gösterilen kurtarma kodları
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFAEnableResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/mfa/disable:
    post:
      tags: [auth]
      summary: MFA kapat (kod doğrula)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MFACodeRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/oauth/{provider}:
    get:
      tags: [auth]
      summary: OAuth/SSO akışını başlat (sağlayıcıya yönlendirir)
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, github] } }
      responses:
        "302": { description: Sağlayıcının yetkilendirme sayfasına yönlendirme }
        "422": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/oauth/{provider}/callback:
    get:
      tags: [auth]
      summary: OAuth/SSO callback (kod → token çifti)
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, github] } }
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Token çifti
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/permissions:
    get:
      tags: [auth]
      summary: Oturum açmış kullanıcının izin listesi
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: İzin listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions:
                    type: array
                    items: { type: string }
  /api/v1/users:
    post:
      tags: [users]
      summary: Yeni kullanıcı oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Oluşturulan kullanıcı (yerelleştirilmiş mesaj + veri)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserViewWithMessage" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    get:
      tags: [users]
      summary: Kullanıcıları listele (sayfa tabanlı sayfalama + filtreleme + sıralama)
      description: |
        Varsayılan: `page`/`limit` offset sayfalama (panel ile uyumlu).
        İsteğe bağlı `cursor` ile keyset sayfalama; yanıtta `next_cursor` dolabilir.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: role, in: query, schema: { type: string }, description: Dinamik rol adı (ör. admin, editor) }
        - { name: active, in: query, schema: { type: boolean } }
        - { name: search, in: query, schema: { type: string } }
        - name: deleted
          in: query
          schema: { type: string, enum: ["", only, all], default: "" }
          description: "Silinme kapsamı: boş = canlı, only = yalnızca silinenler, all = tümü"
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, enum: [100, 500, 1000], default: 100 } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
          description: |
            Keyset imleci (`next_cursor`). Verildiğinde `page` yok sayılır (page=1);
            panel offset sayfalaması etkilenmez.
      responses:
        "200":
          description: Sayfalanmış kullanıcı listesi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPage" }
  /api/v1/users/me:
    get:
      tags: [users]
      summary: Kimliği doğrulanmış kullanıcının profili ve izinleri
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Profil + permissions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeView" }
  /api/v1/users/me/name:
    patch:
      tags: [users]
      summary: Görünen adı güncelle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
      responses:
        "200":
          description: Güncellenmiş profil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserViewWithMessage" }
        "422": { $ref: "#/components/responses/Problem" }
  /api/v1/users/me/email:
    patch:
      tags: [users]
      summary: E-posta adresini güncelle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { $ref: "#/components/schemas/Email" }
      responses:
        "200":
          description: Güncellenmiş profil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserViewWithMessage" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/users/me/phone:
    patch:
      tags: [users]
      summary: Telefon numarasını güncelle veya kaldır
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { $ref: "#/components/schemas/Phone" }
      responses:
        "200":
          description: Güncellenmiş profil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserViewWithMessage" }
        "422": { $ref: "#/components/responses/Problem" }
  /api/v1/users/me/locale:
    patch:
      tags: [users]
      summary: Kalıcı dil tercihini güncelle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locale]
              properties:
                locale:
                  type: string
                  example: en
                  description: BCP-47 temel dil kodu (ör. tr, en)
      responses:
        "200":
          description: Güncellenmiş profil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserViewWithMessage" }
        "422": { $ref: "#/components/responses/Problem" }
  /api/v1/users/me/notifications:
    get:
      tags: [notifications]
      summary: Kullanıcının uygulama içi bildirimleri (sayfa tabanlı sayfalama)
      description: |
        Merkezi bildirim sisteminin uygulama içi (inapp) türü kalıcı olarak
        saklanır. Bu uç kimliği doğrulanmış kullanıcının kendi bildirimlerini
        en yeni önce döner. Bildirim GÖNDERME dışa açık değildir; iç servisler
        ve domain event'leri Dispatcher üzerinden gönderir.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, required: false, schema: { type: integer, enum: [100, 500, 1000], default: 100 } }
      responses:
        "200":
          description: Bildirim sayfası
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPage" }
    delete:
      tags: [notifications]
      summary: Tüm bildirimleri sil
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /api/v1/users/me/notifications/unread-count:
    get:
      tags: [notifications]
      summary: Okunmamış bildirim sayısı
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Okunmamış sayısı
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread: { type: integer, example: 3 }
  /api/v1/users/me/notifications/read-all:
    post:
      tags: [notifications]
      summary: Tüm bildirimleri okundu işaretle
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /api/v1/users/me/notifications/{id}/read:
    post:
      tags: [notifications]
      summary: Bildirimi okundu işaretle
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/users/me/notifications/{id}:
    delete:
      tags: [notifications]
      summary: Bildirimi sil
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/users/{id}:
    get:
      tags: [users]
      summary: Kullanıcı getir
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Kullanıcı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserView" }
        "404": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [users]
      summary: Kullanıcıyı yazılımsal sil (soft delete, yalnızca admin)
      description: |
        Kaydı fiziksel olarak silmez; deleted_at damgalar. Silinen kayıt canlı
        sorgularda görünmez ve e-postası yeni kayıtlarda yeniden kullanılabilir.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/users/{id}/email:
    patch:
      tags: [users]
      summary: E-posta değiştir (kendi hesabı veya users:email:change:any izni)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { $ref: "#/components/schemas/Email" }
      responses:
        "200":
          description: Güncellenmiş kullanıcı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserViewWithMessage" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /api/v1/users/{id}/role:
    patch:
      tags: [users]
      summary: Kullanıcı rolünü değiştir (users:role:change izni; oturumlar iptal edilir)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeRoleRequest" }
      responses:
        "200":
          description: Güncellenmiş kullanıcı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserViewWithMessage" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Problem:
      description: RFC 7807 hata yanıtı
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    LoginRequest:
      type: object
      description: İlk adımda email+password; MFA adımında mfa_challenge+mfa_code gönderilir.
      properties:
        email: { $ref: "#/components/schemas/Email" }
        password: { type: string }
        mfa_code:
          type: string
          description: 6 haneli TOTP veya tek kullanımlık kurtarma kodu
        mfa_challenge:
          type: string
          description: İlk adımın 401 yanıtından alınan, 5 dakika geçerli opak challenge
        turnstile_token: { $ref: "#/components/schemas/TurnstileToken" }
      oneOf:
        - required: [email, password]
        - required: [mfa_challenge, mfa_code]
...
    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
...